Refactor API routes into modular package
